+++ src/main/java/com/zoraw/cinema/model/db/mongo/ScreeningRepository.java
@@ -1,7 +1,7 @@
 package com.zoraw.cinema.model.db.mongo;
 
 import com.zoraw.cinema.model.db.mongo.dao.ScreeningDao;
-import com.zoraw.cinema.model.dto.Screening;
+import com.zoraw.cinema.model.domain.Screening;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
