+++ src/main/java/com/zoraw/cinema/model/db/mongo/mapper/ReservationMapper.java
@@ -1,7 +1,7 @@
 package com.zoraw.cinema.model.db.mongo.mapper;
 
 import com.zoraw.cinema.model.db.mongo.dao.ReservationDao;
-import com.zoraw.cinema.model.dto.Reservation;
+import com.zoraw.cinema.model.domain.Reservation;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
