+++ src/main/java/com/zoraw/cinema/model/db/mongo/mapper/SeatMapper.java
@@ -1,7 +1,7 @@
 package com.zoraw.cinema.model.db.mongo.mapper;
 
 import com.zoraw.cinema.model.db.mongo.dao.SeatDao;
-import com.zoraw.cinema.model.dto.Seat;
+import com.zoraw.cinema.model.domain.Seat;
 import org.mapstruct.Mapper;
 
 import java.util.Set;
