+++ src/main/java/com/zoraw/cinema/model/domain/TicketType.java
@@ -1,4 +1,4 @@
-package com.zoraw.cinema.model.dto;
+package com.zoraw.cinema.model.domain;
 
 public enum TicketType {
     ADULT,
