+++ src/main/java/com/zoraw/cinema/model/service/ReservationService.java
@@ -1,7 +1,7 @@
 package com.zoraw.cinema.model.service;
 
-import com.zoraw.cinema.model.dto.Reservation;
-import com.zoraw.cinema.model.dto.ReservationResponse;
+import com.zoraw.cinema.model.domain.Reservation;
+import com.zoraw.cinema.model.domain.ReservationResponse;
 
 public interface ReservationService {
