+++ src/main/java/com/zoraw/cinema/model/service/impl/ReservationCreationServiceImpl.java
@@ -5,9 +5,9 @@ import com.zoraw.cinema.model.db.mongo.ScreeningRepository;
 import com.zoraw.cinema.model.db.mongo.dao.ScreeningDao;
 import com.zoraw.cinema.model.db.mongo.mapper.ReservationMapper;
 import com.zoraw.cinema.model.db.mongo.mapper.ScreeningMapper;
-import com.zoraw.cinema.model.dto.Reservation;
-import com.zoraw.cinema.model.dto.Room;
-import com.zoraw.cinema.model.dto.Seat;
+import com.zoraw.cinema.model.domain.Reservation;
+import com.zoraw.cinema.model.domain.Room;
+import com.zoraw.cinema.model.domain.Seat;
 import com.zoraw.cinema.model.exception.BusinessException;
 import com.zoraw.cinema.model.service.ReservationCreationService;
 import lombok.RequiredArgsConstructor;
