+++ src/main/java/com/zoraw/cinema/model/service/impl/ScreeningServiceImpl.java
@@ -2,9 +2,8 @@ package com.zoraw.cinema.model.service.impl;
 
 import com.zoraw.cinema.model.db.mongo.ScreeningRepository;
 import com.zoraw.cinema.model.db.mongo.mapper.ScreeningMapper;
-import com.zoraw.cinema.model.dto.Movie;
-import com.zoraw.cinema.model.dto.Screening;
-import com.zoraw.cinema.model.dto.ScreeningBasic;
+import com.zoraw.cinema.model.domain.Movie;
+import com.zoraw.cinema.model.domain.Screening;
 import com.zoraw.cinema.model.service.ScreeningService;
 import com.zoraw.cinema.model.service.dto.MovieWithScreeningsDto;
 import com.zoraw.cinema.model.service.dto.ScreeningDto;
@@ -20,10 +19,6 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 @Service
 class ScreeningServiceImpl implements ScreeningService {
-
-    private static final Comparator<ScreeningBasic> SCREENING_COMPARATOR = Comparator.comparing(
-            (ScreeningBasic screeningBasic) -> screeningBasic.getMovie().getTitle())
-            .thenComparing(ScreeningBasic::getTime);
 
     private final ScreeningRepository screeningRepository;
     private final ScreeningMapper screeningMapper;
