+++ src/test/groovy/com/zoraw/cinema/model/domain/RoomDaoDtoTest.groovy
@@ -1,4 +1,4 @@
-package com.zoraw.cinema.model.dto
+package com.zoraw.cinema.model.domain
 
 import spock.lang.Specification
 import spock.lang.Unroll
